+++ services/user/internal/domain/client_profile.go
@@ -3,12 +3,12 @@ package domain
 import "time"
 
 type ClientProfile struct {
-	AccountID               int
-	BonusBalance            int64
-	BonusCategoryID         *int
-	BonusCategoryExpiresAt  *time.Time
-	BonusExpiresAt          *time.Time
-	StreakCount             int
-	LastOrderDate           *time.Time
-	PremiumExpiresAt        *time.Time
+	AccountID              int64
+	BonusBalance           int64
+	BonusCategoryID        *int64
+	BonusCategoryExpiresAt *time.Time
+	BonusExpiresAt         *time.Time
+	StreakCount            int
+	LastOrderDate          *time.Time
+	PremiumExpiresAt       *time.Time
 }
