+++ shared/pkg/common/context_keys.go
@@ -8,7 +8,7 @@ const RequestIDKey contextKey = "requestID"
 
 const UserIDKey contextKey = "userID"
 
-func GetUserID(ctx context.Context) (int, bool) {
-	id, ok := ctx.Value(UserIDKey).(int)
+func GetUserID(ctx context.Context) (int64, bool) {
+	id, ok := ctx.Value(UserIDKey).(int64)
 	return id, ok
 }
